Challenge 7c
I cannot for the life of me figure out how to get my items to match. I don't know if the Color Matching Behaviour script was edited off camera or what but I can't get it to work for the life of me even after getting hints from other videos. Because of that I can't test my Game Over or Replay.