More challenge work
I'm trying to catch up so I had the Labs done but not the challenges so now I'm just trying to get everything done. I have no idea to fix the instantiating problem and I have no idea how to get the 321Go! Thing to work. the only thing I can think of is to use the scene builder but I know thats not what the assignment wants us to do.