Progress on the challenge
Okay I have fixed the instantiating issue, figured out how to get it to replay, and the fixed the color match issue. Now I just need to figure out the 321Go! thing and how to import audio. I've been successful in the past with audio but for some reason it isn't working this time. Additionally, I keep getting a stack overflow issue that I need to fix. I am starting to lose motivation to complete this challenge because I have probably spent 30 hours just working and reworking everything and creating issues to solve, but I am determined to complete the game before the end of the semester. When I get something right, I really enjoy the coding process, but this learning curve has been exceptionally brutal.